+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Fixed
 - Cut videos more accurately by dropping the no-negative-ts flag from the  cut call to ffmpeg.
+- Show durations in the UI including overrun.
 
 ## [3.25.0] - 2024-02-13
 ### Changed